requirements.txt
# requirements.txt
torch>=1.8.0 # Specify a version compatible with your CUDA/CPU setup if needed
transformers>=4.0.0
numpy>=1.18.0
Pillow>=8.0.0